feature: 메모리 사이즈 소수점 2째자리까지로 변경

+++ core/src/main/java/com/multi/core/dto/AgentInfoDTO.java
@@ -5,5 +5,5 @@ import lombok.Builder;
 import java.io.Serializable;
 
 @Builder
-public record AgentInfoDTO (int cpuCores, long memorySize, String osInfo) implements Serializable {
+public record AgentInfoDTO (Integer cpuCores, Double memorySize, String osInfo) implements Serializable {
 }
