+++ server/src/main/java/com/multi/server/agent/dto/AgentDTO.java
@@ -4,7 +4,7 @@ import com.multi.server.agent.entity.Agent;
 import lombok.Builder;
 
 @Builder
-public record AgentDTO (String agentIp, Integer cpuCores, Long memorySize, String osInfo) {
+public record AgentDTO (String agentIp, Integer cpuCores, Double memorySize, String osInfo) {
     public Agent toEntity() {
         return Agent.builder()
                 .agentIp(agentIp)
