+++ server/src/main/java/com/multi/server/agent/dto/AgentsSearchDTO.java
@@ -3,5 +3,5 @@ package com.multi.server.agent.dto;
 import lombok.Builder;
 
 @Builder
-public record AgentsSearchDTO (String agentIp, Integer cpuCores, Long memorySize, String osInfo) {
+public record AgentsSearchDTO (String agentIp, Integer cpuCores, Double memorySize, String osInfo) {
 }
